use-cases/tiktok_ref_video_gen/pyproject.toml
[project]
name = "tiktok-ref-video-gen"
version = "0.1.0"
description = "TikTok reference video generator using BytePlus AI services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.21.0",
    "veadk-python>=0.5.25",
    "veadk-python[extensions]>=0.5.25",
    "agentkit-sdk-python>=0.5.1",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2024.1.0",
    "tos>=2.6.0",
    "httpx>=0.27.0"
]